requirements.txt
bcrypt>=4.0.1
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
opencv-python>=4.10.0.84
Pillow>=10.4.0
paddlepaddle
paddleocr
customtkinter>=5.2.0

